Extract shared manifest loading.
